Add namespace for CLI
