+++ lib/bundled_gem/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module BundledGem
-  VERSION = "0.0.0"
+  VERSION = "0.0.1"
 end
